Apply Android Studio inspections
* Remove invalid manifest entries
* Remove redundant finals
* Remove redundant semicolons
* Remove unnecessary boxing/unboxing
* Remove redundant type casts
* Remove redundant toString calls
* Replace StringBuilders with Strings
* Remove manual array operations
* Replace StringBuffer with StringBuilder
* Replace for with foreach
* Remove unused try/catch
* Remove unused collections
* Remove empty if statements
* Remove unused return values
* Remove redundant local variables
* Remove redundant assignments and unused variables
* Remove unused logo change
* Remove unused strings
* Fix string comparison

+++ astrid/src/main/java/com/todoroo/astrid/gtasks/GtasksMetadataService.java
@@ -180,8 +180,8 @@ public final class GtasksMetadataService extends SyncMetadataService<GtasksTaskC
      * Gets the remote id string of the previous sibling task
      */
     public String getRemoteSiblingId(String listId, Metadata gtasksMetadata) {
-        final AtomicInteger indentToMatch = new AtomicInteger(gtasksMetadata.getValue(GtasksMetadata.INDENT).intValue());
-        final AtomicLong parentToMatch = new AtomicLong(gtasksMetadata.getValue(GtasksMetadata.PARENT_TASK).longValue());
+        final AtomicInteger indentToMatch = new AtomicInteger(gtasksMetadata.getValue(GtasksMetadata.INDENT));
+        final AtomicLong parentToMatch = new AtomicLong(gtasksMetadata.getValue(GtasksMetadata.PARENT_TASK));
         final AtomicReference<String> sibling = new AtomicReference<String>();
         OrderedListIterator iterator = new OrderedListIterator() {
             @Override
@@ -190,7 +190,7 @@ public final class GtasksMetadataService extends SyncMetadataService<GtasksTaskC
                 if (t == null || t.isDeleted()) {
                     return;
                 }
-                int currIndent = metadata.getValue(GtasksMetadata.INDENT).intValue();
+                int currIndent = metadata.getValue(GtasksMetadata.INDENT);
                 long currParent = metadata.getValue(GtasksMetadata.PARENT_TASK);
 
                 if (currIndent == indentToMatch.get() && currParent == parentToMatch.get()) {
